ChatBot/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
groq>=0.3.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.2
pyjwt>=2.8.0

